Backend/requirements.txt
Flask
Flask-JWT-Extended
Werkzeug
python-dotenv
